--- src/app/Api/Core/Abstract/BaseBuilder.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Api.Core.Abstract
-{
-    public abstract class BaseBuilder<T>
-    {
-        protected T data;
-
-        public BaseBuilder()
-        {
-
-        }
-
-        public T Build()
-        {
-            return data;
-        }
-    }
-}
