--- src/app/Api/Core/Abstract/BaseController.cs
@@ -1,26 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.ModelBinding;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace Api.Core.Abstract
-{
-    [ApiController]
-    [Route("[controller]")]
-    public abstract class BaseController : ControllerBase
-    {
-        protected string CurrentUserId => User.Claims.FirstOrDefault(x => x.Type == "id")?.Value;
-
-        protected virtual object GetErrorsFromModelState(ModelStateDictionary modelState)
-        {
-            Dictionary<string, string> errors = modelState.ToDictionary(x => x.Key, y => y.Value.Errors.FirstOrDefault()?.ErrorMessage);
-
-            return new { errors };
-        }
-
-        protected virtual object GetErrorsModel(params object[] errors)
-        {
-            return new { errors };
-        }
-    }
-}
