--- src/app/Api/Core/Abstract/IAuthService.cs
@@ -1,7 +1,0 @@
-﻿namespace Api.Core.Abstract
-{
-    public interface IAuthService
-    {
-        string CreateAuthToken(string id);
-    }
-}
