+++ src/app/Api/Core/DbViews/BaseBuilder.cs
@@ -1,0 +1,17 @@
+﻿namespace Api.Core.DbViews
+{
+    public abstract class BaseBuilder<T>
+    {
+        protected T data;
+
+        public BaseBuilder()
+        {
+
+        }
+
+        public T Build()
+        {
+            return data;
+        }
+    }
+}
