+++ src/app/Api/Core/DbViews/User/UserFactory.cs
@@ -1,0 +1,17 @@
+﻿namespace Api.Core.DbViews.User
+{
+    public class UserFactory
+    {
+        public User UnverifiedUser()
+        {
+            UserBuilder builder = new UserBuilder();
+            return builder.NotVerifiedEmail().Build();
+        }
+
+        public User VerifiedUser()
+        {
+            UserBuilder builder = new UserBuilder();
+            return builder.Build();
+        }
+    }
+}
