+++ src/app/Api/Core/Interfaces/DAL/IUserRepository.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using Api.Core.DbViews.User;
+
+namespace Api.Core.Interfaces.DAL
+{
+    public interface IUserRepository : IRepository<User>
+    {
+        Task<IEnumerable<User>> GetAllUsers();
+    }
+}
