+++ src/app/Api/Core/Interfaces/Services/App/IAuthService.cs
@@ -1,0 +1,7 @@
+﻿namespace Api.Core.Interfaces.Services.App
+{
+    public interface IAuthService
+    {
+        string CreateAuthToken(string id);
+    }
+}
