+++ src/app/Api/Core/Interfaces/Services/App/IEmailService.cs
@@ -1,0 +1,9 @@
+﻿namespace Api.Core.Interfaces.Services.App
+{
+    public interface IEmailService
+    {
+        void SendSignupWelcome(object data);
+
+        void SendForgotPassword(object data);
+    }
+}
