+++ src/app/Api/Core/Interfaces/Services/Infrastructure/ITokenService.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Api.Core.Interfaces.Services.Infrastructure
+{
+    //public interface ITokenService
+    //{
+    //    List<AccessToken> CreateAuthTokens(string userId);
+    //}
+}
