+++ src/app/Api/Core/Services/Infrastructure/TokenService.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Api.Core.Services.Infrastructure
+{
+    public class TokenService
+    {
+    }
+}
