+++ src/app/Api/Core/Settings/AppConstants.cs
@@ -1,0 +1,10 @@
+﻿namespace Api.Core.Settings
+{
+    public static class AppConstants
+    {
+        public static class DbDocuments
+        {
+            public static string Users = "users";
+        }
+    }
+}
