+++ src/app/Api/Core/Settings/AppSettings.cs
@@ -1,0 +1,8 @@
+﻿namespace Api.Core.Settings
+{
+    public class AppSettings
+    {
+        public string WebUrl { get; set; }
+        public string LandingUrl { get; set; }
+    }
+}
