+++ src/app/Api/Core/Settings/DbSettings.cs
@@ -1,0 +1,8 @@
+﻿namespace Api.Core.Settings
+{
+    public class DbSettings
+    {
+        public string ConnectionString { get; set; }
+        public string Database { get; set; }
+    }
+}
