+++ src/app/Api/Properties/launchSettings.json
@@ -1,5 +1,4 @@
 {
-  "$schema": "http://json.schemastore.org/launchsettings.json",
   "iisSettings": {
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
@@ -8,10 +7,11 @@
       "sslPort": 0
     },
     "iisExpress": {
-      "applicationUrl": "http://localhost:3001/",
+      "applicationUrl": "http://localhost:5001/",
       "sslPort": 0
     }
   },
+  "$schema": "http://json.schemastore.org/launchsettings.json",
   "profiles": {
     "IIS Express": {
       "commandName": "IISExpress",
