--- src/app/Api/Settings/AppConstants.cs
@@ -1,10 +1,0 @@
-﻿namespace Api.Settings
-{
-    public static class AppConstants
-    {
-        public static class DbDocuments
-        {
-            public static string Users = "users";
-        }
-    }
-}
