--- src/app/Api/Settings/AppSettings.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Api.Settings
-{
-    public class AppSettings
-    {
-        public string WebUrl { get; set; }
-        public string LandingUrl { get; set; }
-    }
-}
