--- src/app/Api/Settings/DbSettings.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Api.Settings
-{
-    public class DbSettings
-    {
-        public string ConnectionString { get; set; }
-        public string Database { get; set; }
-    }
-}
